+++ src/markdown/common/consts.ts
@@ -324,6 +324,8 @@ export const terminatedRules = {
 
 export const mathTokenTypes = ["display_math", "inline_math", "equation_math_not_number", "equation_math"];
 
+export const DEFAULT_TYPESET_CACHE_SIZE = 50000;
+
 export const codeHighlightDef = {
   auto: false,
   code: true