+++ src/markdown/mathpix-markdown-plugins.ts
@@ -1,5 +1,6 @@
 import { MarkdownIt } from 'markdown-it';
 import { MathpixMarkdownModel, ParserErrors } from "../mathpix-markdown-model";
+import { DEFAULT_TYPESET_CACHE_SIZE } from './common/consts';
 import { resetTextCounter } from './mdPluginText';
 import { resetTheoremEnvironments } from './md-theorem/helper';
 import { rest_mmd_footnotes_list } from './md-latex-footnotes/utils';
@@ -42,7 +43,8 @@ export const mathpixMarkdownPlugin = (md: MarkdownIt, options) => {
     copyToClipboard = false,
     renderOptions = null,
     previewUuid = "",
-    enableSizeCalculation = false
+    enableSizeCalculation = false,
+    typesetCacheSize = DEFAULT_TYPESET_CACHE_SIZE,
   } = options;
   Object.assign(md.options, smiles);
   Object.assign(md.options, {
@@ -70,7 +72,8 @@ export const mathpixMarkdownPlugin = (md: MarkdownIt, options) => {
     copyToClipboard: copyToClipboard,
     renderOptions: renderOptions,
     previewUuid: previewUuid,
-    enableSizeCalculation: enableSizeCalculation
+    enableSizeCalculation: enableSizeCalculation,
+    typesetCacheSize: typesetCacheSize,
   });
 
   md